requirements.txt
sqlalchemy==2.0.23
python-dotenv==1.0.0
pandas==2.1.3
pyodbc==5.0.1
openpyxl==3.1.2
notebook==7.0.6